moved MainGearLength, NoseGearLength to geometry subsystem from mass subsystem
