+++ aviary/subsystems/mass/flops_based/landing_group.py
@@ -1,11 +1,7 @@
 import openmdao.api as om
 
-from aviary.subsystems.mass.flops_based.landing_gear import (
-    AltLandingGearMass,
-    LandingGearMass,
-    MainGearLength,
-    NoseGearLength,
-)
+from aviary.subsystems.geometry.flops_based.landing_gear import MainGearLength, NoseGearLength
+from aviary.subsystems.mass.flops_based.landing_gear import AltLandingGearMass, LandingGearMass
 from aviary.subsystems.mass.flops_based.landing_mass import LandingMass, LandingTakeoffMassRatio
 from aviary.variable_info.functions import add_aviary_option
 from aviary.variable_info.variables import Aircraft
